+++ Assets/_Scripts/PaulMemes/HealthToMeter.cs
@@ -28,12 +28,12 @@ public class HealthToMeter : MonoBehaviour
         compHealth.MaxHealthChanged -= Health_MaxHealthChanged;
     }
 
-    private void Health_CurrentHealthChanged(float newHealthCurrent)
+    private void Health_CurrentHealthChanged(int newHealthCurrent)
     {    
         compMeter.SetCurrentValue(newHealthCurrent);
     }
 
-    private void Health_MaxHealthChanged(float newHealthMax)
+    private void Health_MaxHealthChanged(int newHealthMax)
     {
         compMeter.SetMaxValue(newHealthMax);
     }
